Switch Dockerfile ARGs to all lower-case.
When ARGs and ENV vars have the same name attempting to override an ENV
var that was set in a previous layer with the ARG fails - it picks up
the ENV var value instead of the arg.

Because of this we are switching over to all lower-case ARG names. This
also helps clearly distinguish between values that are ARGs and those
that are ENV vars.